Move enough_capacity from vector to utility
vector.c is no more needed
