--- src/vector.c
@@ -1,9 +1,0 @@
-#include "vector.h"
-
-size_t enough_capacity(size_t size) {
-  size_t capacity = 1;
-  while (capacity < size) {
-    capacity *= 2;
-  }
-  return capacity;
-}
